main features done

+++ internal/models/objects/property.go
@@ -5,7 +5,7 @@ type Property struct {
 	PropertyTypeId int `json:"property_type_id"`
 	PropertyType string `json:"property_type" `
 	Address string `json:"address"`
-	Price int `json:"price"`
+	Price float64 `json:"price"`
 	Rooms int `json:"rooms"`
 	Area float32 `json:"area"`
 	Description string `json:"description"`
@@ -15,7 +15,7 @@ type Prop struct {
 	Id int `json:"id"`
 	PropertyTypeId int `json:"property_type_id"`
 	AddressId int `json:"address_id"`
-	Price int `json:"price"`
+	Price float64 `json:"price"`
 	Rooms int `json:"rooms"`
 	Area float32 `json:"area"`
 	Description string `json:"description"`
